+++ lib/td_dd/permissions.ex
@@ -5,41 +5,18 @@ defmodule TdDd.Permissions do
 
   alias TdDd.Auth.Claims
 
-  @doc """
-  Check if authenticated user has a permission in any domain.
+  def authorized?(%Claims{jti: jti}, permissions) when is_list(permissions) do
+    TdCache.Permissions.has_any_permission_on_resource_type?(jti, permissions, "domain")
+  end
 
-  ## Examples
-
-      iex> authorized?(%Claims{}, "create")
-      false
-
-  """
   def authorized?(%Claims{jti: jti}, permission) do
     TdCache.Permissions.has_permission?(jti, permission)
   end
 
-  @doc """
-  Check if authenticated user has a permission in a domain.
-
-  ## Examples
-
-      iex> authorized?(%Claims{}, "create", 12)
-      false
-
-  """
   def authorized?(%Claims{jti: jti}, permission, domain_id) do
     TdCache.Permissions.has_permission?(jti, permission, "domain", domain_id)
   end
 
-  @doc """
-  Check if authenticated user has a any permission in a domain.
-
-  ## Examples
-
-      iex> authorized_any?(%Claims{}, [:create, :delete], "business_concept", 12)
-      false
-
-  """
   def authorized_any?(%Claims{jti: jti}, permissions, domain_id) do
     TdCache.Permissions.has_any_permission?(jti, permissions, "domain", domain_id)
   end